+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+Breaking changes!
+- Made package web compatible
+- Allow package to pass cache dir path as `Future<String>?`
+
 ## 1.0.0
 
 Initial release with all basic functionalities.